create `solana-zk-sdk-pod` and migrate encryption pod types
