+++ Cargo.toml
@@ -1,6 +1,6 @@
 [workspace]
 resolver = "2"
-members = ["zk-sdk", "zk-sdk-wasm-js"]
+members = ["zk-sdk", "zk-sdk-pod", "zk-sdk-wasm-js"]
 
 [workspace.package]
 authors = ["Anza Maintainers <maintainers@anza.xyz>"]
@@ -53,6 +53,7 @@ solana-seed-phrase = "3.0.0"
 solana-signature = { version = "3.0.0", default-features = false }
 solana-signer = "3.0.0"
 solana-zk-sdk = { path = "zk-sdk", version = "4.0.0" }
+solana-zk-sdk-pod = { path = "zk-sdk-pod", version = "0.1.0" }
 solana-zk-sdk-wasm-js = { path = "zk-sdk-wasm-js", version = "0.1.0" }
 subtle = "2.6.1"
 thiserror = "2.0.17"
