+++ zk-sdk/src/encryption/pod/elgamal.rs
@@ -1,65 +1,16 @@
 //! Plain Old Data types for the ElGamal encryption scheme.
 
+use solana_zk_sdk_pod::encryption::elgamal::{
+    PodDecryptHandle, PodElGamalCiphertext, PodElGamalPubkey,
+};
 #[cfg(not(target_os = "solana"))]
 use {
     crate::{
         encryption::elgamal::{DecryptHandle, ElGamalCiphertext, ElGamalPubkey},
         errors::ElGamalError,
     },
-    curve25519_dalek::ristretto::CompressedRistretto,
+    // curve25519_dalek::ristretto::CompressedRistretto,
 };
-use {
-    crate::{
-        encryption::{DECRYPT_HANDLE_LEN, ELGAMAL_CIPHERTEXT_LEN, ELGAMAL_PUBKEY_LEN},
-        pod::{impl_from_bytes, impl_from_str},
-    },
-    base64::{prelude::BASE64_STANDARD, Engine},
-    bytemuck::Zeroable,
-    std::fmt,
-};
-
-/// Maximum length of a base64 encoded ElGamal public key
-const ELGAMAL_PUBKEY_MAX_BASE64_LEN: usize = 44;
-
-/// Maximum length of a base64 encoded ElGamal ciphertext
-const ELGAMAL_CIPHERTEXT_MAX_BASE64_LEN: usize = 88;
-
-/// Maximum length of a base64 encoded ElGamal decrypt handle
-const DECRYPT_HANDLE_MAX_BASE64_LEN: usize = 44;
-
-/// The `ElGamalCiphertext` type as a `Pod`.
-#[derive(Clone, Copy, bytemuck_derive::Pod, bytemuck_derive::Zeroable, PartialEq, Eq)]
-#[repr(transparent)]
-pub struct PodElGamalCiphertext(pub(crate) [u8; ELGAMAL_CIPHERTEXT_LEN]);
-
-impl fmt::Debug for PodElGamalCiphertext {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{:?}", self.0)
-    }
-}
-
-impl fmt::Display for PodElGamalCiphertext {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", BASE64_STANDARD.encode(self.0))
-    }
-}
-
-impl Default for PodElGamalCiphertext {
-    fn default() -> Self {
-        Self::zeroed()
-    }
-}
-
-impl_from_str!(
-    TYPE = PodElGamalCiphertext,
-    BYTES_LEN = ELGAMAL_CIPHERTEXT_LEN,
-    BASE64_LEN = ELGAMAL_CIPHERTEXT_MAX_BASE64_LEN
-);
-
-impl_from_bytes!(
-    TYPE = PodElGamalCiphertext,
-    BYTES_LEN = ELGAMAL_CIPHERTEXT_LEN
-);
 
 #[cfg(not(target_os = "solana"))]
 impl From<ElGamalCiphertext> for PodElGamalCiphertext {
@@ -77,31 +28,6 @@ impl TryFrom<PodElGamalCiphertext> for ElGamalCiphertext {
     }
 }
 
-/// The `ElGamalPubkey` type as a `Pod`.
-#[derive(Clone, Copy, Default, bytemuck_derive::Pod, bytemuck_derive::Zeroable, PartialEq, Eq)]
-#[repr(transparent)]
-pub struct PodElGamalPubkey(pub(crate) [u8; ELGAMAL_PUBKEY_LEN]);
-
-impl fmt::Debug for PodElGamalPubkey {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{:?}", self.0)
-    }
-}
-
-impl fmt::Display for PodElGamalPubkey {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", BASE64_STANDARD.encode(self.0))
-    }
-}
-
-impl_from_str!(
-    TYPE = PodElGamalPubkey,
-    BYTES_LEN = ELGAMAL_PUBKEY_LEN,
-    BASE64_LEN = ELGAMAL_PUBKEY_MAX_BASE64_LEN
-);
-
-impl_from_bytes!(TYPE = PodElGamalPubkey, BYTES_LEN = ELGAMAL_PUBKEY_LEN);
-
 #[cfg(not(target_os = "solana"))]
 impl From<ElGamalPubkey> for PodElGamalPubkey {
     fn from(decoded_pubkey: ElGamalPubkey) -> Self {
@@ -118,17 +44,6 @@ impl TryFrom<PodElGamalPubkey> for ElGamalPubkey {
     }
 }
 
-/// The `DecryptHandle` type as a `Pod`.
-#[derive(Clone, Copy, Default, bytemuck_derive::Pod, bytemuck_derive::Zeroable, PartialEq, Eq)]
-#[repr(transparent)]
-pub struct PodDecryptHandle(pub(crate) [u8; DECRYPT_HANDLE_LEN]);
-
-impl fmt::Debug for PodDecryptHandle {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{:?}", self.0)
-    }
-}
-
 #[cfg(not(target_os = "solana"))]
 impl From<DecryptHandle> for PodDecryptHandle {
     fn from(decoded_handle: DecryptHandle) -> Self {
@@ -136,13 +51,13 @@ impl From<DecryptHandle> for PodDecryptHandle {
     }
 }
 
-// For proof verification, interpret pod::DecryptHandle as CompressedRistretto
-#[cfg(not(target_os = "solana"))]
-impl From<PodDecryptHandle> for CompressedRistretto {
-    fn from(pod_handle: PodDecryptHandle) -> Self {
-        Self(pod_handle.0)
-    }
-}
+// // For proof verification, interpret pod::DecryptHandle as CompressedRistretto
+// #[cfg(not(target_os = "solana"))]
+// impl From<PodDecryptHandle> for CompressedRistretto {
+//     fn from(pod_handle: PodDecryptHandle) -> Self {
+//         Self(pod_handle.0)
+//     }
+// }
 
 #[cfg(not(target_os = "solana"))]
 impl TryFrom<PodDecryptHandle> for DecryptHandle {
@@ -153,19 +68,11 @@ impl TryFrom<PodDecryptHandle> for DecryptHandle {
     }
 }
 
-impl fmt::Display for PodDecryptHandle {
-    fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
-        write!(f, "{}", BASE64_STANDARD.encode(self.0))
-    }
-}
-
-impl_from_str!(
-    TYPE = PodDecryptHandle,
-    BYTES_LEN = DECRYPT_HANDLE_LEN,
-    BASE64_LEN = DECRYPT_HANDLE_MAX_BASE64_LEN
-);
-
-impl_from_bytes!(TYPE = PodDecryptHandle, BYTES_LEN = DECRYPT_HANDLE_LEN);
+// impl fmt::Display for PodDecryptHandle {
+//     fn fmt(&self, f: &mut fmt::Formatter) -> fmt::Result {
+//         write!(f, "{}", BASE64_STANDARD.encode(self.0))
+//     }
+// }
 
 #[cfg(test)]
 mod tests {
