+++ zk-sdk/src/sigma_proofs/batched_grouped_ciphertext_validity/handles_3.rs
@@ -158,15 +158,12 @@ mod test {
     use {
         super::*,
         crate::{
-            encryption::{
-                elgamal::ElGamalKeypair,
-                pedersen::Pedersen,
-                pod::{
-                    elgamal::{PodDecryptHandle, PodElGamalPubkey},
-                    pedersen::PodPedersenCommitment,
-                },
-            },
+            encryption::{elgamal::ElGamalKeypair, pedersen::Pedersen},
             sigma_proofs::pod::PodBatchedGroupedCiphertext3HandlesValidityProof,
+        },
+        solana_zk_sdk_pod::encryption::{
+            elgamal::{PodDecryptHandle, PodElGamalPubkey},
+            pedersen::PodPedersenCommitment,
         },
         std::str::FromStr,
     };
