+++ zk-sdk/src/sigma_proofs/grouped_ciphertext_validity/handles_2.rs
@@ -248,15 +248,12 @@ mod test {
     use {
         super::*,
         crate::{
-            encryption::{
-                elgamal::ElGamalKeypair,
-                pedersen::Pedersen,
-                pod::{
-                    elgamal::{PodDecryptHandle, PodElGamalPubkey},
-                    pedersen::PodPedersenCommitment,
-                },
-            },
+            encryption::{elgamal::ElGamalKeypair, pedersen::Pedersen},
             sigma_proofs::pod::PodGroupedCiphertext2HandlesValidityProof,
+        },
+        solana_zk_sdk_pod::encryption::{
+            elgamal::{PodDecryptHandle, PodElGamalPubkey},
+            pedersen::PodPedersenCommitment,
         },
         std::str::FromStr,
     };
