+++ zk-sdk/src/sigma_proofs/grouped_ciphertext_validity/handles_3.rs
@@ -279,15 +279,12 @@ mod test {
     use {
         super::*,
         crate::{
-            encryption::{
-                elgamal::ElGamalKeypair,
-                pedersen::Pedersen,
-                pod::{
-                    elgamal::{PodDecryptHandle, PodElGamalCiphertext, PodElGamalPubkey},
-                    pedersen::PodPedersenCommitment,
-                },
-            },
+            encryption::{elgamal::ElGamalKeypair, pedersen::Pedersen},
             sigma_proofs::pod::PodGroupedCiphertext3HandlesValidityProof,
+        },
+        solana_zk_sdk_pod::encryption::{
+            elgamal::{PodDecryptHandle, PodElGamalCiphertext, PodElGamalPubkey},
+            pedersen::PodPedersenCommitment,
         },
         std::str::FromStr,
     };
