+++ zk-sdk/src/zk_elgamal_proof_program/proof_data/batched_grouped_ciphertext_validity/handles_2.rs
@@ -5,16 +5,6 @@
 //! grouped-ciphertext validity proof is shorter and more efficient than two individual
 //! grouped-ciphertext validity proofs.
 
-use {
-    crate::{
-        encryption::pod::{
-            elgamal::PodElGamalPubkey, grouped_elgamal::PodGroupedElGamalCiphertext2Handles,
-        },
-        sigma_proofs::pod::PodBatchedGroupedCiphertext2HandlesValidityProof,
-        zk_elgamal_proof_program::proof_data::{ProofType, ZkProofData},
-    },
-    bytemuck_derive::{Pod, Zeroable},
-};
 #[cfg(not(target_os = "solana"))]
 use {
     crate::{
@@ -27,6 +17,16 @@ use {
     },
     bytemuck::bytes_of,
     merlin::Transcript,
+};
+use {
+    crate::{
+        sigma_proofs::pod::PodBatchedGroupedCiphertext2HandlesValidityProof,
+        zk_elgamal_proof_program::proof_data::{ProofType, ZkProofData},
+    },
+    bytemuck_derive::{Pod, Zeroable},
+    solana_zk_sdk_pod::encryption::{
+        elgamal::PodElGamalPubkey, grouped_elgamal::PodGroupedElGamalCiphertext2Handles,
+    },
 };
 
 /// The instruction data that is needed for the
