+++ zk-sdk/src/zk_elgamal_proof_program/proof_data/percentage_with_cap.rs
@@ -19,12 +19,12 @@ use {
 };
 use {
     crate::{
-        encryption::pod::pedersen::PodPedersenCommitment,
         pod::PodU64,
         sigma_proofs::pod::PodPercentageWithCapProof,
         zk_elgamal_proof_program::proof_data::{ProofType, ZkProofData},
     },
     bytemuck_derive::{Pod, Zeroable},
+    solana_zk_sdk_pod::encryption::pedersen::PodPedersenCommitment,
 };
 
 /// The instruction data that is needed for the `ProofInstruction::VerifyPercentageWithCap`
